Koala Image
I replicated this image from the one I saw on medium.com. I didn't follow the tutorial though. I just used the end-result as a guideline to recreate the image. Other than the colors used and the end-result, both the HTML and CSS code used are very different. 

Here is the link to the medium webpage:  https://medium.com/coding-artist/a-beginners-guide-to-pure-css-images-ef9a5d069dd2